+++ src/content/Settings/Settings.css
@@ -5,3 +5,20 @@
     text-align: left;
     padding-left: 5px;
 }
+
+.SettingsButton {
+    width: 50%;
+    padding: 7px 20px;
+    font-size: 12px;
+    color: #fff;
+    background-color: #007bff;;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s, box-shadow 0.3s;
+    margin: 10px 10px 2px 10px;
+}
+
+.SettingsButton:hover {
+    background-color: #0068d6;;
+}
